refactor: reduce -t / --tag to a flag
BREAKING CHANGE: the `-t` (`--tag`) option used do and be used exactly like the
`-b` option but added a commit-and-tag to the operation. So `versync -b major`
would bump the major version. And `versync -t major` would bump the major and
commit-and-tag. The `-t` option is now a plain on/off flag. Where you used to do
`versync -t major`, you now have to do `versync -b major -t` to turn on the
commit-and-tag operation *in addition* to the version bump.