feat: add support for `-b sync`
`-b sync` allows propagating a version value from package.json to the other
source files provided to versync.

BREAKING CHANGE:

This change required changes to the API. If you only used ``Runner`` and called
``run`` on it, there should be no change that affects you. If you used the other
methods of ``Runner`` directly, or used the utility functions that are exported
by versync, you may need to update your code. The changes are:

- ``VersionInfo`` now includes a ``source`` field which is the file that
  contains the version number. This is not "breaking" but may be useful for you.

- The exported function ``verify``:

  * It no longer takes an ``expectedVersion`` parameter. Instead of checking
    that the files have an expected version, it checks that the files are
    consistent.

  * It used to return an array of strings with the names of the problematic
    source files. Now it returns an object whose ``consistent`` field indicates
    whether the version numbers are consistent. The ``versions`` field is an
    array of ``VersionInfo`` elements.