docs: add a missing breaking change
BREAKING CHANGE:

The API was changed to remove the ``verify`` flag from
``RunnerOptions``. Previously you could use it to specify a run that would not
verify anything by setting the flag to ``false``. It would have an effect only
when not bumping and amounted to tell versync to do nothing at all, which is not
terribly useful. Removing the flag simplifies the code. Now calling run
will *always* result in versync performing at least a verification of the
source.