feat: add support for running `git add` by itself
versync has supported `--tag` for checking it the modified sources and creating
a tag. However, when using versync together with `npm version` or a similar tool
that takes care of doing commits and setting tags, it is not possible to use
`tag`. At the same time, it is desirable to have versync be able to `git add` the
files it modifies. We've added flags for this purpose.