feat: versync now uses execFile rather than exec to launch children
exec launches a shell. It is generally wasteful to launch a shell, and it may
lead to bizarre results because the parameters passed to git are subject to
interpretation by the shell.

BREAKING CHANGE: the switch from exec to execFile *could* in theory cause some
usages of versync that previously worked to now fail. I'd expect the likelihood
of such occurence to be extremely small because attempts to take advantage of
the shell interpretation provided by exec would most likely fail in other
ways. (E.g. putting quotes in a version number would make semver checks fail.)